Move render test to a separate file
